requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bandit==1.7.5
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
gitdb==4.0.10
GitPython==3.1.31
markdown-it-py==2.2.0
mdurl==0.1.2
pbr==5.11.1
Pygments==2.14.0
pytz==2022.7.1
PyYAML==6.0
rich==13.3.2
smmap==5.0.0
sqlparse==0.4.3
stevedore==5.0.0
typing-extensions==4.5.0
